requirements.txt
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
python-docx>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
ollama>=0.1.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
